getopt replaced by ezOptionParser-0.2.2
Command line option handling now used ezOptionParser-0.2.2. The
benefit is that it alows creation and use of configure files to avoid
lengthy command lines. Options on command line take precedence, then
options specified in --read-config filename, then options specified in
the file "config" if that exists in the launch folder.  Disadvantage is that
it does not allow options of the form --cell-permute=2 but requres space
separation between option name and argument.